Fix warnings

+++ roommedia/main.cpp
@@ -553,9 +553,9 @@ static void selectTrack(int track) {
     play();
 
     GstFormat fmt = GST_FORMAT_TIME;
-    long int len;
+    gint64 len;
     if (gst_element_query_duration (pipeline, &fmt, &len))
-        g_print("total %ld\n",len/1000000);
+        g_print("total %ld\n",(long int)len/1000000);
 }
 
 static void volume(gdouble volume) {
@@ -586,14 +586,14 @@ static void position(int position) {
 
 static gboolean getduration(gpointer) {
     GstFormat fmt = GST_FORMAT_TIME;
-    long int pos, len;
+    gint64 pos, len;
 
     if (gst_element_query_position (pipeline, &fmt, &pos)) {
-        g_print("current %ld\n",pos/1000000);
+        g_print("current %ld\n",(long int)pos/1000000);
     } else
         g_print("current 0\n");
     if (gst_element_query_duration (pipeline, &fmt, &len)) {
-        g_print("total %ld\n",len/1000000);
+        g_print("total %ld\n",(long int)len/1000000);
     } else
         g_print("total 0\n");
     return false;
